refactor: 파일 구조를 변경함에 따른 리펙토링(#19)
