+++ src/main/java/demago/gamjappang/global/security/jwt/JwtTokenProvider.java
@@ -1,13 +1,11 @@
 package demago.gamjappang.global.security.jwt;
 
-import demago.gamjappang.user.adapter.out.persistence.UserJpaEntity;
 import demago.gamjappang.global.error.exception.GamjaException;
-import demago.gamjappang.user.application.port.out.TokenPort;
+import demago.gamjappang.user.applicationcore.port.out.TokenPort;
 import demago.gamjappang.user.domain.model.Role;
 import demago.gamjappang.user.domain.model.User;
 import demago.gamjappang.user.exception.JwtErrorCode;
 import demago.gamjappang.global.security.userdetails.UserPrincipal;
-import demago.gamjappang.global.security.userdetails.UserPrincipalService;
 import io.jsonwebtoken.*;
 import io.jsonwebtoken.security.Keys;
 import java.nio.charset.StandardCharsets;
